remove printlns
